requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# AI/ML services
openai==1.35.10

# Transport stack
httpx==0.27.2
httpcore==1.0.5

# Web utilities
python-multipart==0.0.6
jinja2==3.1.2

# Configuration and utilities
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0

# Audio processing
pydub==0.25.1

# HTML parsing (for stream detection)
beautifulsoup4==4.12.2

# Date/time handling
pytz==2023.3

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1

# Optional audio system dependencies (uncomment if needed)
# sounddevice==0.4.6
# soundfile==0.12.1
# numpy==1.24.3

# Note: sqlite3, threading, logging, subprocess, pathlib, json are built into Python
